Add RPC API and conductor manager for traits
Adds two new methods to the conductor RPC API, add_node_traits and
remove_node_traits, and provides implementations for them in the
conductor manager. add_node_traits can be used to add one or more traits
to a node, or to replace all the traits for a node.  remove_node_traits
can be used to remove one or more traits from a node, or to remove all
traits from a node.

The conductor RPC API version is bumped to 1.44.

Change-Id: I0181df6a41e603874677246066d84bf4ac4f433c
Partial-Bug: #1722194