Modifying 'whole_disk_image_url' and 'whole_disk_image_checksum' variable
Need to set variable whole_disk_image_url and
whole_disk_image_checksum in tempest.conf
irrespective of IPXE is enabled or disabled.

Also setting the ping_timeout for standalone ironic
like we are setting ssh_timeout in case node is
taking more time for becoming active.

This change is required for iLO-CI as iLO-CI is
running tests as standalone ironic.

Change-Id: Idef1bc876fe440824894332a211af9a10401e49c
Closes-Bug: #1748849