Only set default network interface flat if enabled in config
This patch updates the driver factory logic which sets a default for the
network interface when there is not one set in the config file. Before
this patch the logic forces "flat" as the default network interface for
all hardware types and classic drivers even if flat isn't enabled in the
config file, resulting in a misleading error.

This patch changes the code to only set "flat" as the default for
classic drivers if its enabled in the enabled_network_interfaces
configuration option in the config file, and to not override a default
for hardware types at all because the default for hardware types should
either come from the config file or from the hardware type's
supported_network_interfaces list.

Change-Id: Ia8676d3483ddc78df8766dc1baaf2db6b5686050
Closes-Bug: #1744332