Switch the CI to hardware types
Note that the job names are not changed to reduce the diff of this
patch. They will be renamed in later changes.

For the direct deploy, copy Swift temporary URL configuration from
devstack-gate (it hardcoded agent_* there).

Also clean up no-op conditions on drivers left from the times when
the job definitions were parametrized.

Change-Id: Ib0297f8936bba9449a02fb1a878981ab55687ef9
Partial-Bug: #1690185