Fix unit tests for UEFI iSCSI boot for ILO
https://review.openstack.org/468288 added some unit tests for the
UEFI iSCSI boot for ILO, which fail when proliantutils is installed
in the virtualenv (it was not in the gate).

The reason for this is that they do not pass a value to the constructor
of ilo_error.IloError and ilo_error.IloCommandNotSupportedInBiosError,
which makes the unit tests fail, as seen when the package was built
in RDO [1].

This patch adds the required value, as used by other unit tests.

[1] - https://trunk.rdoproject.org/centos7/73/71/737147706405785b3879135872d6efb2953a9aea_0520bb51

Change-Id: I50d919047ddb9edf1a1ae1a994a8f71e420718d9