Explicitly mark earliest-version for release notes
This uses reno's `earliest-version` directive to explicitly point out
where the release notes for a given branch should begin. There is a bug
in reno when dealing with stable branches, where it does not always find
the correct version to begin with.

It also updates our release docs to indicate we should continue marking
this. This step can be removed when the reno bug is fixed, though it
doesn't hurt to continue marking this explicitly.

Change-Id: I6502ff95a52c2c855356e9875291f27ec44e7ffa
Related-Bug: #1746076