Remove mode argument from boot.(prepare|clean_up)_ramdisk
Ideally, the boot interface shouldn't care if it's booting an image for
deploy or rescue. The first step to unwinding this is not passing the
mode argument into the boot interface - for now, we infer it from the
state. Also stop validating whether the boot interface methods have a
mode argument, as this is totally broken anyway (due to the decorator on
the method). The rest of the boot interface's knowledge about deploy vs
rescue can be eliminated in the future, as we shouldn't rework too much
of that during feature freeze.

We fix the bug in validation this way for now, for two reasons:

* This argument really doesn't belong, and it's only been on master for
  six days. Get it out before people use it.
* Library updates are currently frozen; fixing the decorator isn't an
  option right now.

Change-Id: Icdeb870e9fd9cf836ff7ab97624189c8436adaba
Closes-Bug: #1746730