Ensure we unbind flat network ports and clear BM mac addresses
This patch fixes a bug in the ironic nova neutron interaction triangle,
where a port in a flat network can be left with bound and with the mac
address of a baremetal server if it is not deleted after a deployment or
deployment failure. The fix is to ensure that the network interfaces
reset the mac address to a random mac address when it unbinds the port
in the unconfigure step for provisioning and tenant networks, and then
rebind and set the mac address for the configure steps.

Co-Authored-By: Harald Jensås <hjensas@redhat.com>
Story: #2004428
Task: #28087
Change-Id: I11fff92e0a58ac68e795c003c14a336bceba6d89
(cherry picked from commit db16984b47afe014d2eadd32b56ccf1c347fb53d)