Use zuul.override_checkout  instead of custom branch_override var
To override the branch correctly in both zuulv3's playbook loading we
have to use override_checkout in the job. This variable already gives us
the information we need to set the override branch, so use that instead
of the custom branch_override variable so we don't have to define it
twice in the job.

Change-Id: Iec7be918f75e3400197748385b287fb945d761ac