Fix rare HTTP 400 from port list API
This may happen when a node is deleted in parallel with calling
the port list API. Ports are fetched, then we try do fetch their
nodes and port groups. If either of them are removed in the meantime,
the API fails with HTTP 400. This change works around it.

Change-Id: Ie2d4c46c031ee86976abb6107433cdde87a4345a
Closes-Bug: #1748893
(cherry picked from commit 52dcc642d372e23fd59be44e0f9f5627fac5cec4)