Register traits on nodes in devstack
This change adds support for assiging traits to nodes in the ironic
devstack plugin.  The set of traits added is configured by
IRONIC_DEFAULT_TRAITS, which defaults to a single trait - CUSTOM_GOLD

Using nodes with traits during testing ensures that we are exercising
node traits code paths in the Ironic CI.

Node traits are not used for upgrade testing, as the queens devstack
plugin does not currently register traits. Once this change is
backported to queens, we can enable traits during upgrade testing.

Change-Id: I18b9dbf8a5b16e36e0cc14f290b12ffdd634364a
Related-Bug: #1722194
Related-Bug: #1750027