[devstack] block iPXE boot from HTTPS TempURLs
While generally supported by iPXE, in DevStack we do not have any
facilities in place to boot iPXE from a HTTPS URL when the server is
using a custom certificate.
This is moslty concerning ability to boot iPXE directly from Swift's
TempURL when SSL is enabled on Swift in DevStack.

Just die with proper message when IRONIC_IPXE_USE_SWIFT and tls-proxy
are both enabled.

Change-Id: Ie0e01828f05b2706d310d8110e6e8dd8c4a03e4e