tox.ini: flake8: Remove I202 from ignore list
Remove I202 (Additional newline in a group of imports) from the ignore
list. I202 is a check from the flake8-import-order [1] package.

There was a note to remove this check once the tempest plugin was
removed. As the tempest plugin has been removed we can now remove the
exclusion of this check.

Update files which failed this test.

[1] https://github.com/PyCQA/flake8-import-order

Change-Id: I6018d5a90174b3d4b6b8d8e05cee1c1104ccfde9