Add descriptions for config option choices
oslo.config 5.2.0 changed some of the options
(oslo_config.cfg.StrOpt & PortOpt) so that
the choices parameter will now accept a sequence
of tuples, where each tuple is of the form
(choice, description).

If a description is not provided, the sample
config file is generated with
'<No description provided>' next to the choices.

This adds descriptions to the choices.

Change-Id: I7e32ff8e86d2760af49fd088188b94d19ca76e84
Closes-Bug: #1755914