Allow Swift endpoint override
Swiftclient uses public endpoint by default. Ironic uses the base URL
from Swift connection to build TempURLs for generated images.
Some drivers (e.g. iLO) use those TempURLs to mount images as vmedia.
With public URLs it will fail if the BMC doesn't have access to the
public network.

This change introduces an option to explicitly set the endpoint URL
used for Swift.

This is a stable-only change as the problem is fixed by refactoring
changes in later releases. This is the only version where Ironic uses
public Swift endpoints and there is no option to override this.

Change-Id: I639a421fa06fff7ab07b8eab557531b8f36c5ed9
Closes-Bug: #1755164
Related-Bug: #1699547