Rework Bare Metal service overview in the install guide
Currently we have a detailed explanation of how ironic is architected,
but it resides in the user guide for some reason. This change moves
relevant bits to the install guide to the existing "Bare Metal service
overview" section. The text is cleaned up (e.g. s/ironic/Bare Metal/)
and ordered to simplify reading. A link to ironic-ui is added.

The user guide still contains the bits that are helpful for users
to understand the bare metal deployment process.

Change-Id: I17e38454f7893c00b71ac29328bd80d3978ed1f4