Power fault recovery: apply fault
This patch implements setting and using the fault field.

For each case currently maintenance is set to True, the fault is set
accordingly. A periodic task is added to check power state for nodes
in maintenance due to power failure, maintenance is cleared if the
power state of a node can be retrieved.

When a node is taken out of maintenance by user, the fault is
cleared (if there is any).

Story: #1596107
Task: #10469

Change-Id: Ic4ab20af9022a2d06bdac567e7a098f3ba08570a
Partial-Bug: #1596107