Stop overriding tempdir in unit test
It may causes failure in test__prepare_variables_configdrive_file to
use a directory which doesn't exist as tempdir paramter because
interfaces using ipmitool see if the directory exists when they are
loaded. Actually, this test may fail since ipmitool vendor interface
is enabled.

This patch fixes this test to use the default tempdir.

Change-Id: I685a2aa0dcfe919b7e5318eb00f070ced6cad812
Story: #2001755
Task: #12153