Resolve pep8 E402 errors and no longer ignore E402
Fix the pep8 E402 (Module level import not at top of file) errors.

For the Alembic files move the definitions of 'revision' and
'down_revision' below the imports. This is now done in Alembic [1]

For other files fix as needed by moving the imports or lines before
the imports.

In a few cases add a "# noqa E402" to whitelist the issue.

[1] https://bitbucket.org/zzzeek/alembic/commits/dc2aae0374b76ccab8494be6e103d84089bc449b

Change-Id: I48c96d5da0cb747b3ca3fceea9b4ffa85a9ebe22