Run jobs if requirements change
The dsvm jobs were not being run if only test-requirements.txt was
modified. However, devstack installs test-requirements so we should
run those jobs if that file changes.

The grenade multinode jobs were not being run if requirements.txt
was modified -- this was due to an extra '|' which caused it to be
considered an irrelevant file :-(

Change-Id: I1a49915ca52a93e77849dc5ead655656ad1bb34e