Remove ip parameter from ipxe command line
Due to dracut changes, we can no longer use BOOTIF on the
same command line, and the lack of an ``autoconf`` parameter
results in DHCP being utilized, which is also incompatible with
the ``ip`` command on the kernel command line.

The history of the ip parameter is to allow the kernel to skip
executing DHCP again, but typially it is done anyway so it seemes
that it was redundant to begin with.

Change-Id: I1d2ebafea39d27941178f22b36dcc37a73848cf3
Story: #2001969
Task: #16141