Remove endpoint_type from configuration
python-swiftclient stopped supporting the temp url structure used when radosgw
was set as the endpoint_type in ocata, meaning only Newton and older versions
of python-swiftclient will work.  Newton is deprecated, so remove the option.

This breaks the deprecation cycle, but since it has been not working for so
long it needs to just be dropped.

Change-Id: Ibdc93b049b7e1ae34cac9e1f599786439c46a685