Disable periodic tasks if interval set to 0
For periodic tasks that are specified with the decorator @perodics.periodic(),
a ValueError exception was raised if a value <= 0 was specified for any of the
spacing values (taken from configuration options).

Specifying a value <=0 used to work, to disable the task altogether. It broke
when we switched to using the futurist package (some time in mitaka cycle).

This fixes it so that setting these configuration options to 0 (or a negative
value) will disable the periodic tasks:
- [conductor]sync_power_state_interval: sync power states for the nodes
- [conductor]check_provision_state_interval:
  - check deployments and time out if the deployment takes too long
  - check the status of cleaning a node and time out if it takes too long
  - check the status of inspecting a node and time out if it takes too long
  - check for and handle nodes that are taken over by new conductors (if an
    old conductor disappeared)
- [conductor]send_sensor_data_interval: send sensor data to ceilometer
- [conductor]sync_local_state_interval: refresh a conductor's copy of the
  consistent hash ring. If any mappings have changed, determines which,
  if any, nodes need to be "taken over". The ensuing actions could include
  preparing a PXE environment, updating the DHCP server, and so on.
- [oneview]periodic_check_interval:
  - check for nodes taken over by OneView users
  - check for nodes freed by OneView users

Change-Id: I62708e239295344d0dcf0bff7dd68ec8c34ab9a0
Story: #2002059
Task: #19708
(cherry picked from commit f39aae0c95567d78e9a29dbb7e74abe1104c97d0)