Fix error when deleting a non-existent port
Fixes an issue seen during node tear down where a port being deleted by
the Bare Metal service could be deleted by the Compute service, leading
to an unhandled error from the Networking service. This change simply
ignores the PortNotFoundClient exception raises, as is done when
attempting to unbind a port that does not exist. See [1] for details.

[1] https://storyboard.openstack.org/#!/story/2002637>

Change-Id: Icd2aaa1276e75f08f99553bbc9a873102d896fb9
Story: 2002637
Task: 22285