During cleaning, use current node.driver_internal_info
When iterating over the clean steps, the code in the conductor
keeps a copy of the node's driver_internal_info and uses that
to update the node's copy. However, it is possible that the
node's driver_internal_info is modified after the copy is made
(e.g. if a driver's clean step method modifies/saves it).
This patch changes it to use the current node's driver_internal_info.

Story: #2002688
Task: #22513
Change-Id: I92ad67b58343e5c6c1cca916269f134a5c78dcc6