include all versions of Node in release_mappings
For a particular ironic release, we need to include all the
possible versions of a particular object. Node (so far) has
two versions: 1.24 and 1.25 in this current release.

(Without the two versions, there could be issues upgrading
ironic from an older-master release with Node version 1.24,
to a newer-master release with Node version 1.25.)

Change-Id: I27c8d12add23ad72ea2debd1aec2bf6afaff2f90