Change PXE logic to always link macs with UEFI
Our logic for elilo and grub excluded the MAC
address based links from being created, which are
handy for stand-alone use cases.

This patch changes the default to always write
MAC address based links, and allows the interface
to tolerate an exception being raised if the
``[dhcp]dhcp_provider`` is set to ``none``.

This effectively fixes the issue where users that
do not use neutron were unable to use UEFI.

Story: #2002887
Task: #22848
Change-Id: Ib8b9b2e3281184cd4e8d9a1a33a012f93edf237c