Don't handle warnings as errors
At some point between changes to the doc builds in the
gates and local doc builds, warnings were no longer being
treated as errors. In that time, multiple changes have occured
with sphinx and related tooling that results in errors being
generatd out of the results from autodoc, as our program model
uses classes which are inherited in various modules and that
contain docstrings.

Due to dulicate method names in the api automatic doc genration
we raise a warning. Guidence from the error is to add :noindex:
to one of the duplicates to suppress the error, however those
text entries in the index are being autogenerated, so that is
not an option.

Since there seems to be no simple path to take to remedy this,
removing the -W flag from the docs build to enable developers
to simply execute `tox` to test their changes

Story: 2003020
Change-Id: I4ae8dee9a624781b231659fb24c5c4d53353da3f