Ignore bashate E044
Bashate introduces a new error, E044, which attempts to
identify any questionable conditionals that would be
incorrectly using bracketing. That being said, it is
modeled on keeping the design simple, and errors on more
condensed lines where conditionals are included.

Since we don't need this, and the case where it is erroring
is invalid as the test is overly simple, lets ignore it.

Change-Id: Ie46dcd6cd6100f3c7e35c2c817828e45bd6b7921