Fix for failure in cleaning in iRMC driver
This patch will fix an issue for BIOS configuration
during combine out-of-band an in-band in manual clean.

The cleaning operation will be fail, if an in-band clean step were to
execute after the completion of out-of-band clean step that
performs reboot of the node. This fixing is following other patch [1]
from Shivanand.

[1] https://review.openstack.org/#/c/580136/

Tested on TX2540 M1 Fujitsu servers successfully.
Co-authored-By: Dao Cong Tien <tiendc@vn.fujitsu.com>

Change-Id: I0b37b93851cc56141c2fe78793296c50ad25f9ad