A vision
During the rocky PTG, we discussed the vision
for ironic. We learned to use the same words.
We reached the same page.

As a result, the week felt very productive.

And so, we should actually write some of this down.

Accordingly the patch proposes a draft vision, as well
as sets the context of the discussions in an attept to
have a point or document that we can refer back to at
a later point in time.

The contents are based off of the Rocky ironic PTG vision
etherpad[1].

[1]: https://etherpad.openstack.org/p/ironic-rocky-ptg-vision

Change-Id: Ie7208cc5aeade96409b658041b9edb4976512ad7