Use max version of an object
In the code for handling rolling upgrades (and pinned ironic versions),
we were assuming that for an object class (eg 'Node') that had more than
one version in a release, the first version was the latest one. This was
error-prone as humans don't always know/remember to put the latest version
as the first one in the list.

This patch removes the assumption and gets the maximum version in the list.

Change-Id: Ie0a789f8bfa6f286b58cc7211be85ee18b2800f5